Se corrige error de pantalla de CostDetail

+++ src/Sofco.Core/Models/ManagementReport/CostDetailModel.cs
@@ -20,7 +20,7 @@ namespace Sofco.Core.Models.ManagementReport
         public string Display { get; set; }
         public float? Value { get; set; }
         public float? OriginalValue { get; set; }
-        public int? Adjustment { get; set; }
+        public float? Adjustment { get; set; }
         public DateTime MonthYear { get; set; }
         public int CostDetailId { get; set; }
         public bool HasAlocation { get; set; }
